opea-docker/agent/requirements.txt
requests
duckduckgo_search
html2tex